+++ Scenes/Units/tank_gen_2.tscn
@@ -205,7 +205,7 @@ alpha_curve = SubResource("CurveTexture_hmto5")
 
 [sub_resource type="SphereMesh" id="SphereMesh_uqx8m"]
 
-[node name="TankGen2" type="CharacterBody3D" node_paths=PackedStringArray("Death", "PrimaryWeaponSystem", "LODManager", "_healthSystem") groups=["damageable", "units"]]
+[node name="TankGen2" type="CharacterBody3D" node_paths=PackedStringArray("Death", "PrimaryWeaponSystem", "_healthSystem") groups=["damageable", "units"]]
 transform = Transform3D(-1, 0, -8.74228e-08, 0, 1, 0, 8.74228e-08, 0, -1, 0, 0, 0)
 script = ExtResource("1_eb2nj")
 VehicleType = 2
@@ -217,8 +217,11 @@ Acceleration = 8
 Death = NodePath("Death")
 MiniMapRadius = 5.0
 PrimaryWeaponSystem = NodePath("Systems/WeaponSystem")
-LODManager = NodePath("Systems/LODManager")
 _healthSystem = NodePath("Systems/HealthSystem")
+_highId = 4
+_lowId = 3
+PrimaryTurretPath = NodePath("Model/Rig/Turret")
+PrimaryMuzzleContainerPath = NodePath("Model/Rig/Turret/Muzzle")
 metadata/_edit_group_ = true
 
 [node name="Model" parent="." instance=ExtResource("2_smbrw")]
@@ -321,11 +324,6 @@ draw_pass_1 = SubResource("SphereMesh_uqx8m")
 
 [node name="LODManager" type="Node" parent="Systems"]
 script = ExtResource("14_sd0dm")
-_lodNear = 40.0
-_highId = 4
-_lowId = 3
-PrimaryTurretPath = NodePath("../../Model/Rig/Turret")
-PrimaryMuzzleContainerPath = NodePath("../../Model/Rig/Turret/Muzzle")
 
 [node name="CombatSystem" type="Node" parent="Systems" node_paths=PackedStringArray("_unit", "_weaponSystem", "_muzzleFlashParticles", "_attackSound")]
 script = ExtResource("5_k7ata")
