+++ Scripts/Units/CombatSystem.cs
@@ -53,9 +53,9 @@ public partial class CombatSystem : Node
 
 		// Subscribe to LOD sockets (primary vs secondary)
 		if (_slot == WeaponSlot.Primary)
-			_unit.LODManager.SocketsChangedPrimary += OnSocketsChangedList;
+			_unit.SocketsChangedPrimary += OnSocketsChangedList;
 		else
-			_unit.LODManager.SocketsChangedSecondary += OnSocketsChangedList;
+			_unit.SocketsChangedSecondary += OnSocketsChangedList;
 
 		// Wait until first sockets signal before doing anything
 		_socketsDirty = true;
@@ -63,11 +63,10 @@ public partial class CombatSystem : Node
 
 	public override void _ExitTree()
 	{
-		if (_unit?.LODManager == null) return;
 		if (_slot == WeaponSlot.Primary)
-			_unit.LODManager.SocketsChangedPrimary -= OnSocketsChangedList;
+			_unit.SocketsChangedPrimary -= OnSocketsChangedList;
 		else
-			_unit.LODManager.SocketsChangedSecondary -= OnSocketsChangedList;
+			_unit.SocketsChangedSecondary -= OnSocketsChangedList;
 	}
 
 	public override void _PhysicsProcess(double delta)
